requirements.txt
Flask
gunicorn
openai
pymongo
twilio
google-api-python-client
google-auth
google-auth-oauthlib
pytz
pandas
mysql.connector
# fcntl
redis
celery
redlock-py